Enum rusoto_organizations::LeaveOrganizationError

pub enum LeaveOrganizationError {
    AWSOrganizationsNotInUse(String),
    AccessDenied(String),
    AccountNotFound(String),
    ConcurrentModification(String),
    ConstraintViolation(String),
    InvalidInput(String),
    MasterCannotLeaveOrganization(String),
    Service(String),
    TooManyRequests(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by LeaveOrganization

Variants
AWSOrganizationsNotInUse(String)

Your account is not a member of an organization. To make this request, you must use the credentials of an account that belongs to an organization.

AccessDenied(String)

You don't have permissions to perform the requested operation. The user or role that is making the request must have at least one IAM permissions policy attached that grants the required permissions. For more information, see Access Management in the IAM User Guide.

AccountNotFound(String)

We can't find an AWS account with the AccountId that you specified, or the account whose credentials you used to make this request is not a member of an organization.

ConcurrentModification(String)

The target of the operation is currently being modified by a different request. Try again later.

ConstraintViolation(String)

Performing this operation violates a minimum or maximum value limit. For example, attempting to removing the last SCP from an OU or root, inviting or creating too many accounts to the organization, or attaching too many policies to an account, OU, or root. This exception includes a reason that contains additional information about the violated limit:

ACCOUNT_NUMBER_LIMIT_EXCEEDED: You attempted to exceed the limit on the number of accounts in an organization. If you need more accounts, contact AWS Support to request an increase in your limit.

Or, The number of invitations that you tried to send would cause you to exceed the limit of accounts in your organization. Send fewer invitations, or contact AWS Support to request an increase in the number of accounts.

Note: deleted and closed accounts still count toward your limit.

HANDSHAKE_RATE_LIMIT_EXCEEDED: You attempted to exceed the number of handshakes you can send in one day.

OU_NUMBER_LIMIT_EXCEEDED: You attempted to exceed the number of organizational units you can have in an organization.

OU_DEPTH_LIMIT_EXCEEDED: You attempted to create an organizational unit tree that is too many levels deep.

POLICY_NUMBER_LIMIT_EXCEEDED. You attempted to exceed the number of policies that you can have in an organization.

MAX_POLICY_TYPE_ATTACHMENT_LIMIT_EXCEEDED: You attempted to exceed the number of policies of a certain type that can be attached to an entity at one time.

MIN_POLICY_TYPE_ATTACHMENT_LIMIT_EXCEEDED: You attempted to detach a policy from an entity that would cause the entity to have fewer than the minimum number of policies of a certain type required.

ACCOUNT_CANNOT_LEAVE_WITHOUT_EULA: You attempted to remove an account from the organization that does not yet have enough information to exist as a stand-alone account. This account requires you to first agree to the End-User License Agreement (EULA).

ACCOUNT_CANNOT_LEAVE_WITHOUT_PHONE_VERIFICATION: You attempted to remove an account from the organization that does not yet have enough information to exist as a stand-alone account. This account requires you to first complete phone verification.

MASTER_ACCOUNT_PAYMENT_INSTRUMENT_REQUIRED: To create an organization with this account, you first must associate a payment instrument, such as a credit card, with the account.

MEMBER_ACCOUNT_PAYMENT_INSTRUMENT_REQUIRED: To complete this operation with this member account, you first must associate a payment instrument, such as a credit card, with the account.

ACCOUNT_CREATION_RATE_LIMIT_EXCEEDED: You attempted to exceed the number of accounts that you can create in one day.

MASTER_ACCOUNT_ADDRESS_DOES_NOT_MATCH_MARKETPLACE: To create an account in this organization, you first must migrate the organization's master account to the marketplace that corresponds to the master account's address. For example, accounts with India addresses must be associated with the AISPL marketplace. All accounts in an organization must be associated with the same marketplace.

InvalidInput(String)

The requested operation failed because you provided invalid values for one or more of the request parameters. This exception includes a reason that contains additional information about the violated limit:

INVALID_PARTY_TYPE_TARGET: You specified the wrong type of entity (account, organization, or email) as a party.

INVALID_SYNTAX_ORGANIZATION_ARN: You specified an invalid ARN for the organization.

INVALID_SYNTAX_POLICY_ID: You specified an invalid policy ID.

INVALID_ENUM: You specified a value that is not valid for that parameter.

INVALID_FULL_NAME_TARGET: You specified a full name that contains invalid characters.

INVALID_LIST_MEMBER: You provided a list to a parameter that contains at least one invalid value.

MAX_LENGTH_EXCEEDED: You provided a string parameter that is longer than allowed.

MAX_VALUE_EXCEEDED: You provided a numeric parameter that has a larger value than allowed.

MIN_LENGTH_EXCEEDED: You provided a string parameter that is shorter than allowed.

MIN_VALUE_EXCEEDED: You provided a numeric parameter that has a smaller value than allowed.

IMMUTABLE_POLICY: You specified a policy that is managed by AWS and cannot be modified.

INVALID_PATTERN: You provided a value that doesn't match the required pattern.

INVALID_PATTERN_TARGET_ID: You specified a policy target ID that doesn't match the required pattern.

INPUT_REQUIRED: You must include a value for all required parameters.

INVALID_PAGINATION_TOKEN: Get the value for the NextToken parameter from the response to a previous call of the operation.

MAX_FILTER_LIMIT_EXCEEDED: You can specify only one filter parameter for the operation.

MOVING_ACCOUNT_BETWEEN_DIFFERENT_ROOTS: You can move an account only between entities in the same root.

MasterCannotLeaveOrganization(String)

You can't remove a master account from an organization. If you want the master account to become a member account in another organization, you must first delete the current organization of the master account.

Service(String)

AWS Organizations can't complete your request because of an internal service error. Try again later.

TooManyRequests(String)

You've sent too many requests in too short a period of time. The limit helps protect against denial-of-service attacks. Try again later.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl LeaveOrganizationError

fn from_body(body: &str) -> LeaveOrganizationError