Enum rusoto_organizations::ListCreateAccountStatusError

pub enum ListCreateAccountStatusError {
    AWSOrganizationsNotInUse(String),
    AccessDenied(String),
    InvalidInput(String),
    Service(String),
    TooManyRequests(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by ListCreateAccountStatus

Variants
AWSOrganizationsNotInUse(String)

Your account is not a member of an organization. To make this request, you must use the credentials of an account that belongs to an organization.

AccessDenied(String)

You don't have permissions to perform the requested operation. The user or role that is making the request must have at least one IAM permissions policy attached that grants the required permissions. For more information, see Access Management in the IAM User Guide.

InvalidInput(String)

The requested operation failed because you provided invalid values for one or more of the request parameters. This exception includes a reason that contains additional information about the violated limit:

Some of the reasons in the following list might not be applicable to this specific API or operation:

IMMUTABLEPOLICY: You specified a policy that is managed by AWS and cannot be modified.

INPUTREQUIRED: You must include a value for all required parameters.

INVALIDENUM: You specified a value that is not valid for that parameter.

INVALIDFULLNAMETARGET: You specified a full name that contains invalid characters.

INVALIDLISTMEMBER: You provided a list to a parameter that contains at least one invalid value.

INVALIDPARTYTYPETARGET: You specified the wrong type of entity (account, organization, or email) as a party.

INVALIDPAGINATIONTOKEN: Get the value for the NextToken parameter from the response to a previous call of the operation.

INVALIDPATTERN: You provided a value that doesn't match the required pattern.

INVALIDPATTERNTARGETID: You specified a policy target ID that doesn't match the required pattern.

INVALIDROLENAME: You provided a role name that is not valid. A role name can’t begin with the reserved prefix 'AWSServiceRoleFor'.

INVALIDSYNTAXORGANIZATIONARN: You specified an invalid ARN for the organization.

INVALIDSYNTAXPOLICYID: You specified an invalid policy ID.

MAXFILTERLIMITEXCEEDED: You can specify only one filter parameter for the operation.

MAXLENGTHEXCEEDED: You provided a string parameter that is longer than allowed.

MAXVALUEEXCEEDED: You provided a numeric parameter that has a larger value than allowed.

MINLENGTHEXCEEDED: You provided a string parameter that is shorter than allowed.

MINVALUEEXCEEDED: You provided a numeric parameter that has a smaller value than allowed.

MOVINGACCOUNTBETWEENDIFFERENTROOTS: You can move an account only between entities in the same root.

Service(String)

AWS Organizations can't complete your request because of an internal service error. Try again later.

TooManyRequests(String)

You've sent too many requests in too short a period of time. The limit helps protect against denial-of-service attacks. Try again later.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl ListCreateAccountStatusError

pub fn from_body(body: &str) -> ListCreateAccountStatusError