Enum rusoto_organizations::ListHandshakesForOrganizationError

pub enum ListHandshakesForOrganizationError {
    AWSOrganizationsNotInUse(String),
    AccessDenied(String),
    ConcurrentModification(String),
    InvalidInput(String),
    Service(String),
    TooManyRequests(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by ListHandshakesForOrganization

Variants
AWSOrganizationsNotInUse(String)

Your account is not a member of an organization. To make this request, you must use the credentials of an account that belongs to an organization.

AccessDenied(String)

You don't have permissions to perform the requested operation. The user or role that is making the request must have at least one IAM permissions policy attached that grants the required permissions. For more information, see Access Management in the IAM User Guide.

ConcurrentModification(String)

The target of the operation is currently being modified by a different request. Try again later.

InvalidInput(String)

The requested operation failed because you provided invalid values for one or more of the request parameters. This exception includes a reason that contains additional information about the violated limit:

INVALID_PARTY_TYPE_TARGET: You specified the wrong type of entity (account, organization, or email) as a party.

INVALID_SYNTAX_ORGANIZATION_ARN: You specified an invalid ARN for the organization.

INVALID_SYNTAX_POLICY_ID: You specified an invalid policy ID.

INVALID_ENUM: You specified a value that is not valid for that parameter.

INVALID_FULL_NAME_TARGET: You specified a full name that contains invalid characters.

INVALID_LIST_MEMBER: You provided a list to a parameter that contains at least one invalid value.

MAX_LENGTH_EXCEEDED: You provided a string parameter that is longer than allowed.

MAX_VALUE_EXCEEDED: You provided a numeric parameter that has a larger value than allowed.

MIN_LENGTH_EXCEEDED: You provided a string parameter that is shorter than allowed.

MIN_VALUE_EXCEEDED: You provided a numeric parameter that has a smaller value than allowed.

IMMUTABLE_POLICY: You specified a policy that is managed by AWS and cannot be modified.

INVALID_PATTERN: You provided a value that doesn't match the required pattern.

INVALID_PATTERN_TARGET_ID: You specified a policy target ID that doesn't match the required pattern.

INPUT_REQUIRED: You must include a value for all required parameters.

INVALID_PAGINATION_TOKEN: Get the value for the NextToken parameter from the response to a previous call of the operation.

MAX_FILTER_LIMIT_EXCEEDED: You can specify only one filter parameter for the operation.

MOVING_ACCOUNT_BETWEEN_DIFFERENT_ROOTS: You can move an account only between entities in the same root.

Service(String)

AWS Organizations can't complete your request because of an internal service error. Try again later.

TooManyRequests(String)

You've sent too many requests in too short a period of time. The limit helps protect against denial-of-service attacks. Try again later.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl ListHandshakesForOrganizationError

fn from_body(body: &str) -> ListHandshakesForOrganizationError