Struct rusoto_organizations::Organization

pub struct Organization {
    pub arn: Option<String>,
    pub available_policy_types: Option<Vec<PolicyTypeSummary>>,
    pub feature_set: Option<String>,
    pub id: Option<String>,
    pub master_account_arn: Option<String>,
    pub master_account_email: Option<String>,
    pub master_account_id: Option<String>,
}

Contains details about an organization. An organization is a collection of accounts that are centrally managed together using consolidated billing, organized hierarchically with organizational units (OUs), and controlled with policies .

Fields
arn: Option<String>

The Amazon Resource Name (ARN) of an organization.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Organizations User Guide.

available_policy_types: Option<Vec<PolicyTypeSummary>>

A list of policy types that are enabled for this organization. For example, if your organization has all features enabled, then service control policies (SCPs) are included in the list.

Even if a policy type is shown as available in the organization, you can separately enable and disable them at the root level by using EnablePolicyType and DisablePolicyType. Use ListRoots to see the status of a policy type in that root.

feature_set: Option<String>

Specifies the functionality that currently is available to the organization. If set to "ALL", then all features are enabled and policies can be applied to accounts in the organization. If set to "CONSOLIDATED_BILLING", then only consolidated billing functionality is available. For more information, see Enabling All Features in Your Organization in the AWS Organizations User Guide.

id: Option<String>

The unique identifier (ID) of an organization.

The regex pattern for an organization ID string requires "o-" followed by from 10 to 32 lower-case letters or digits.

master_account_arn: Option<String>

The Amazon Resource Name (ARN) of the account that is designated as the master account for the organization.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Organizations User Guide.

master_account_email: Option<String>

The email address that is associated with the AWS account that is designated as the master account for the organization.

master_account_id: Option<String>

The unique identifier (ID) of the master account of an organization.

The regex pattern for an account ID string requires exactly 12 digits.