Struct rusoto_organizations::OrganizationsClient

pub struct OrganizationsClient { /* fields omitted */ }

A client for the Organizations API.

Methods

impl OrganizationsClient

pub fn new(region: Region) -> OrganizationsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> OrganizationsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,