Crate rusoto_organizations

AWS Organizations

If you're using the service, you're probably looking for OrganizationsClient and Organizations.

Structs

AcceptHandshakeRequest

AcceptHandshakeResponse

Account
Contains information about an AWS account that is a member of an organization.

AttachPolicyRequest

CancelHandshakeRequest

CancelHandshakeResponse

Child
Contains a list of child entities, either OUs or accounts.

CreateAccountRequest

CreateAccountResponse

CreateAccountStatus
Contains the status about a CreateAccount request to create an AWS account in an organization.

CreateOrganizationRequest

CreateOrganizationResponse

CreateOrganizationalUnitRequest

CreateOrganizationalUnitResponse

CreatePolicyRequest

CreatePolicyResponse

DeclineHandshakeRequest

DeclineHandshakeResponse

DeleteOrganizationalUnitRequest

DeletePolicyRequest

DescribeAccountRequest

DescribeAccountResponse

DescribeCreateAccountStatusRequest

DescribeCreateAccountStatusResponse

DescribeHandshakeRequest

DescribeHandshakeResponse

DescribeOrganizationResponse

DescribeOrganizationalUnitRequest

DescribeOrganizationalUnitResponse

DescribePolicyRequest

DescribePolicyResponse

DetachPolicyRequest

DisablePolicyTypeRequest

DisablePolicyTypeResponse

EnableAllFeaturesRequest

EnableAllFeaturesResponse

EnablePolicyTypeRequest

EnablePolicyTypeResponse

Handshake
Contains information that must be exchanged to securely establish a relationship between two accounts (an originator and a recipient). For example, when a master account (the originator) invites another account (the recipient) to join its organization, the two accounts exchange information as a series of handshake requests and responses.

HandshakeFilter
Specifies the criteria that are used to select the handshakes for the operation.

HandshakeParty
Identifies a participant in a handshake.

HandshakeResource
Contains additional data that is needed to process a handshake.

InviteAccountToOrganizationRequest

InviteAccountToOrganizationResponse

ListAccountsForParentRequest

ListAccountsForParentResponse

ListAccountsRequest

ListAccountsResponse

ListChildrenRequest

ListChildrenResponse

ListCreateAccountStatusRequest

ListCreateAccountStatusResponse

ListHandshakesForAccountRequest

ListHandshakesForAccountResponse

ListHandshakesForOrganizationRequest

ListHandshakesForOrganizationResponse

ListOrganizationalUnitsForParentRequest

ListOrganizationalUnitsForParentResponse

ListParentsRequest

ListParentsResponse

ListPoliciesForTargetRequest

ListPoliciesForTargetResponse

ListPoliciesRequest

ListPoliciesResponse

ListRootsRequest

ListRootsResponse

ListTargetsForPolicyRequest

ListTargetsForPolicyResponse

MoveAccountRequest

Organization
Contains details about an organization. An organization is a collection of accounts that are centrally managed together using consolidated billing, organized hierarchically with organizational units (OUs), and controlled with policies .

OrganizationalUnit
Contains details about an organizational unit (OU). An OU is a container of AWS accounts within a root of an organization. Policies that are attached to an OU apply to all accounts contained in that OU and in any child OUs.

OrganizationsClient
A client for the Organizations API.

Parent
Contains information about either a root or an organizational unit (OU) that can contain OUs or accounts in an organization.

Policy
Contains rules to be applied to the affected accounts. Policies can be attached directly to accounts, or to roots and OUs to affect all accounts in those hierarchies.

PolicySummary
Contains information about a policy, but does not include the content. To see the content of a policy, see DescribePolicy.

PolicyTargetSummary
Contains information about a root, OU, or account that a policy is attached to.

PolicyTypeSummary
Contains information about a policy type and its status in the associated root.

RemoveAccountFromOrganizationRequest

Root
Contains details about a root. A root is a top-level parent node in the hierarchy of an organization that can contain organizational units (OUs) and accounts. Every root contains every AWS account in the organization. Each root enables the accounts to be organized in a different way and to have different policy types enabled for use in that root.

UpdateOrganizationalUnitRequest

UpdateOrganizationalUnitResponse

UpdatePolicyRequest

UpdatePolicyResponse

Enums

AcceptHandshakeError
Errors returned by AcceptHandshake

AttachPolicyError
Errors returned by AttachPolicy

CancelHandshakeError
Errors returned by CancelHandshake

CreateAccountError
Errors returned by CreateAccount

CreateOrganizationError
Errors returned by CreateOrganization

CreateOrganizationalUnitError
Errors returned by CreateOrganizationalUnit

CreatePolicyError
Errors returned by CreatePolicy

DeclineHandshakeError
Errors returned by DeclineHandshake

DeleteOrganizationError
Errors returned by DeleteOrganization

DeleteOrganizationalUnitError
Errors returned by DeleteOrganizationalUnit

DeletePolicyError
Errors returned by DeletePolicy

DescribeAccountError
Errors returned by DescribeAccount

DescribeCreateAccountStatusError
Errors returned by DescribeCreateAccountStatus

DescribeHandshakeError
Errors returned by DescribeHandshake

DescribeOrganizationError
Errors returned by DescribeOrganization

DescribeOrganizationalUnitError
Errors returned by DescribeOrganizationalUnit

DescribePolicyError
Errors returned by DescribePolicy

DetachPolicyError
Errors returned by DetachPolicy

DisablePolicyTypeError
Errors returned by DisablePolicyType

EnableAllFeaturesError
Errors returned by EnableAllFeatures

EnablePolicyTypeError
Errors returned by EnablePolicyType

InviteAccountToOrganizationError
Errors returned by InviteAccountToOrganization

LeaveOrganizationError
Errors returned by LeaveOrganization

ListAccountsError
Errors returned by ListAccounts

ListAccountsForParentError
Errors returned by ListAccountsForParent

ListChildrenError
Errors returned by ListChildren

ListCreateAccountStatusError
Errors returned by ListCreateAccountStatus

ListHandshakesForAccountError
Errors returned by ListHandshakesForAccount

ListHandshakesForOrganizationError
Errors returned by ListHandshakesForOrganization

ListOrganizationalUnitsForParentError
Errors returned by ListOrganizationalUnitsForParent

ListParentsError
Errors returned by ListParents

ListPoliciesError
Errors returned by ListPolicies

ListPoliciesForTargetError
Errors returned by ListPoliciesForTarget

ListRootsError
Errors returned by ListRoots

ListTargetsForPolicyError
Errors returned by ListTargetsForPolicy

MoveAccountError
Errors returned by MoveAccount

RemoveAccountFromOrganizationError
Errors returned by RemoveAccountFromOrganization

UpdateOrganizationalUnitError
Errors returned by UpdateOrganizationalUnit

UpdatePolicyError
Errors returned by UpdatePolicy

Traits

Organizations
Trait representing the capabilities of the Organizations API. Organizations clients implement this trait.

Type Definitions

AccountArn

AccountId

AccountJoinedMethod

AccountName

AccountStatus

Accounts

ActionType

AwsManagedPolicy

ChildId

ChildType

Children

ConstraintViolationExceptionReason

CreateAccountFailureReason

CreateAccountRequestId

CreateAccountState

CreateAccountStates

CreateAccountStatuses

Email

ExceptionMessage

ExceptionType

GenericArn

HandshakeArn

HandshakeConstraintViolationExceptionReason

HandshakeId

HandshakeNotes

HandshakeParties

HandshakePartyId

HandshakePartyType

HandshakeResourceType

HandshakeResourceValue

HandshakeResources

HandshakeState

Handshakes

IAMUserAccessToBilling

InvalidInputExceptionReason

MaxResults

NextToken

OrganizationArn

OrganizationFeatureSet

OrganizationId

OrganizationalUnitArn

OrganizationalUnitId

OrganizationalUnitName

OrganizationalUnits

ParentId

ParentType

Parents

Policies

PolicyArn

PolicyContent

PolicyDescription

PolicyId

PolicyName

PolicyTargetId

PolicyTargets

PolicyType

PolicyTypeStatus

PolicyTypes

RoleName

RootArn

RootId

RootName

Roots

TargetName

TargetType

Timestamp